# docs/swagger.yaml
basePath: /api
definitions:
  models.Channel:
    properties:
      created_at:
        type: string
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      name:
        type: string
      position:
        type: integer
      server:
        $ref: '#/definitions/models.Server'
      server_id:
        type: integer
      topic:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.Message:
    properties:
      author:
        $ref: '#/definitions/models.User'
      author_id:
        type: integer
      channel:
        $ref: '#/definitions/models.Channel'
      channel_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Server:
    properties:
      channels:
        items:
          $ref: '#/definitions/models.Channel'
        type: array
      created_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/models.ServerMember'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.ServerMember:
    properties:
      created_at:
        type: string
      id:
        type: integer
      joined_at:
        type: string
      nickname:
        type: string
      role:
        type: string
      server:
        $ref: '#/definitions/models.Server'
      server_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      servers:
        items:
          $ref: '#/definitions/models.ServerMember'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  services.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  services.CreateChannelInput:
    properties:
      name:
        type: string
      server_id:
        type: integer
      topic:
        type: string
      type:
        type: string
    required:
    - name
    - server_id
    type: object
  services.CreateServerInput:
    properties:
      icon_url:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  services.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  services.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  services.UpdateChannelInput:
    properties:
      name:
        type: string
      topic:
        type: string
      type:
        type: string
    type: object
  services.UpdateServerInput:
    properties:
      icon_url:
        type: string
      name:
        type: string
    type: object
  services.UpdateUserInput:
    properties:
      avatar_url:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@kinetic.com
    name: API Support
    url: http://localhost:8080/support
  description: Backend API for Kinetic communication platform
  termsOfService: http://localhost:8080/terms
  title: Kinetic API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends password reset email to user
      parameters:
      - description: Email address
        in: body
        name: email
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            type: string
      summary: Forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/services.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AuthResponse'
      summary: Login user
      tags:
      - auth
  /auth/qr-code:
    get:
      description: Generates QR code for 2FA setup
      produces:
      - application/json
      responses:
        "200":
          description: QR code generated
          schema:
            type: string
      summary: Generate QR code
      tags:
      - auth
  /auth/qr-verify:
    post:
      consumes:
      - application/json
      description: Verifies QR code for 2FA
      parameters:
      - description: QR verification code
        in: body
        name: code
        required: true
        schema:
          properties:
            code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: QR verification successful
          schema:
            type: string
      summary: Verify QR code
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refreshes JWT access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed
          schema:
            type: string
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email, username, and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/services.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AuthResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets user password using reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: reset
        required: true
        schema:
          properties:
            password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            type: string
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verifies user email address
      parameters:
      - description: Verification token
        in: body
        name: token
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Email verified
          schema:
            type: string
      summary: Verify email
      tags:
      - auth
  /channels:
    post:
      consumes:
      - application/json
      description: Creates a new channel in a server
      parameters:
      - description: Channel data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/services.CreateChannelInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
      security:
      - BearerAuth: []
      summary: Create a new channel
      tags:
      - channels
  /channels/{id}:
    delete:
      description: Deletes a channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Channel deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete channel
      tags:
      - channels
    get:
      description: Returns a channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
      security:
      - BearerAuth: []
      summary: Get channel by ID
      tags:
      - channels
    patch:
      consumes:
      - application/json
      description: Updates a channel's properties
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Channel update data
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/services.UpdateChannelInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
      security:
      - BearerAuth: []
      summary: Update channel
      tags:
      - channels
  /channels/{id}/messages:
    get:
      description: Returns messages from a channel with pagination
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
      security:
      - BearerAuth: []
      summary: Get channel messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Creates a new message in a channel
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          properties:
            channel_id:
              type: integer
            content:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: Create a new message
      tags:
      - messages
  /channels/{id}/messages/{messageId}:
    delete:
      description: Deletes a message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Message deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a message
      tags:
      - messages
  /servers:
    get:
      description: Returns all servers the authenticated user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Server'
            type: array
      security:
      - BearerAuth: []
      summary: Get user servers
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: Creates a new server and adds the user as owner
      parameters:
      - description: Server data
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/services.CreateServerInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
      security:
      - BearerAuth: []
      summary: Create a new server
      tags:
      - servers
  /servers/{id}:
    delete:
      description: Deletes a server (owner only)
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Server deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete server
      tags:
      - servers
    get:
      description: Returns a server by its ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
      security:
      - BearerAuth: []
      summary: Get server by ID
      tags:
      - servers
    patch:
      consumes:
      - application/json
      description: Updates a server's properties
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Server update data
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/services.UpdateServerInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
      security:
      - BearerAuth: []
      summary: Update server
      tags:
      - servers
  /servers/{id}/channels:
    get:
      description: Returns all channels in a server
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Channel'
            type: array
      security:
      - BearerAuth: []
      summary: Get server channels
      tags:
      - channels
  /servers/{id}/leave:
    post:
      description: Removes the user from a server
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Left server successfully
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Leave server
      tags:
      - servers
  /servers/join:
    post:
      consumes:
      - application/json
      description: Joins a server using invite code
      parameters:
      - description: Invite code
        in: body
        name: invite_code
        required: true
        schema:
          properties:
            invite_code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
      security:
      - BearerAuth: []
      summary: Join server
      tags:
      - servers
  /users/{id}:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile
      parameters:
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /users/{id}/settings:
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's settings
      parameters:
      - description: User settings
        in: body
        name: settings
        required: true
        schema:
          properties:
            language:
              type: string
            theme:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Settings updated
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"